+++ src/Symfony/Component/Serializer/Tests/Fixtures/Attributes/GroupDummy.php
@@ -28,6 +28,12 @@ class GroupDummy extends GroupDummyParent implements GroupDummyInterface
     private $fooBar;
     private $symfony;
 
+    #[Groups(['Default'])]
+    private $default;
+
+    #[Groups(['GroupDummy'])]
+    private $className;
+
     #[Groups(['b'])]
     public function setBar($bar)
     {
@@ -69,6 +75,26 @@ class GroupDummy extends GroupDummyParent implements GroupDummyInterface
     public function getSymfony()
     {
         return $this->symfony;
+    }
+
+    public function setDefault($default)
+    {
+        $this->default = $default;
+    }
+
+    public function getDefault()
+    {
+        return $this->default;
+    }
+
+    public function setClassName($className)
+    {
+        $this->className = $className;
+    }
+
+    public function getClassName()
+    {
+        return $this->className;
     }
 
     public function getQuux()